% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_count2tpm.R
\name{func_count2tpm}
\alias{func_count2tpm}
\title{Convert Count to TPM}
\usage{
func_count2tpm(
  count_data,
  annotation_gtf_dir = "./extdata/gencode.v22.annotation.gtf.gz"
)
}
\arguments{
\item{count_data}{The raw count reads from high-throughput transcriptome sequencing must be in their original state and should not be subjected to log2 transformation or any other normalization processes.}

\item{annotation_gtf_dir}{The path to the annotation file that includes transcript lengths must ensure consistency with the sequencing batch and platform version of the count data. For example, the version number corresponding to the count data at Xena-GDC is version 22 (this function provides the annotation file for version 22 by default).
\itemize{
\item The channel for obtaining the annotation file is the "Annotation Files" section on the following website: https://gdc.cancer.gov/about-data/gdc-data-processing/gdc-reference-files.
}}
}
\value{
The completed TPM (Transcripts Per Million) normalized expression matrix.
}
\description{
Convert the raw read counts from high-throughput transcriptome sequencing to transcripts per million (TPM) normalized data.
\itemize{
\item The function includes a package loading process, utilizing the CRAN mirror from the University of Science and Technology of China for network optimization.
}
}
\examples{
donttest{}
}
