% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_generate_log_system.R
\name{func_generate_log_system}
\alias{func_generate_log_system}
\title{generate a log system for your project}
\usage{
func_generate_log_system()
}
\value{
none.
}
\description{
generate a log system for your project
}
\examples{
donttest{}
}
