% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_RMarkdown.R
\name{func_generate_RMarkdown}
\alias{func_generate_RMarkdown}
\title{Generate RMarkdown.}
\usage{
func_generate_RMarkdown(type, target_path, overwrite = FALSE)
}
\arguments{
\item{type}{Specified RMarkdown styles, should be one of the "downcute", "html_clean", "meterial", "readthedown", and "robobook".}

\item{target_path}{file path}

\item{overwrite}{Do you want to overwrite?}
}
\value{
file path
}
\description{
Generate the main R script for this project, which includes the global configuration and basic package installation of this project. It is a necessary script for the start of the project.
}
\examples{
donttest{}
}
