% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_getPalettes_continuous.R
\name{func_getPalettes_continuous}
\alias{func_getPalettes_continuous}
\title{Obtain a continuous color palette.}
\usage{
func_getPalettes_continuous(
  index_floor = "blue3",
  index_middle = "grey",
  index_celling = "red2",
  color_floor = -10,
  color_middle = 0,
  color_celling = 10
)
}
\arguments{
\item{index_floor}{The lower boundary of the color palette's numerical range.}

\item{index_middle}{The middle boundary of the color palette's numerical range.}

\item{index_celling}{The upper boundary of the color palette's numerical range.}

\item{color_floor}{The color corresponding to the lower boundary of the color palette.}

\item{color_middle}{The color corresponding to the middle boundary of the color palette.}

\item{color_celling}{The color corresponding to the upper boundary of the color palette.}
}
\value{
It returns a function which accepts a vector of numeric values and returns interpolated colors.
}
\description{
Based on the RColorBrewer and circlize packages, conveniently obtain a continuous color palette for a range of numerical values.
\itemize{
\item The function includes a package loading process, utilizing the CRAN mirror from the University of Science and Technology of China for network optimization.
}
}
\examples{
donttest{}
}
