% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_install.R
\name{func_install}
\alias{func_install}
\title{install packages}
\usage{
func_install(
  pkg_list,
  repos = c(CRAN = "https://mirrors.ustc.edu.cn/CRAN/"),
  BioC_mirror = "https://mirrors.ustc.edu.cn/bioc/",
  update = FALSE
)
}
\arguments{
\item{pkg_list}{pkg_list should be a List which includes CRAN, biocManager, or github vecter conteining packages names.}

\item{repos}{repos.}

\item{BioC_mirror}{biocManager repos.}

\item{update}{defult false.}
}
\value{
means nothing
}
\description{
install packages
}
\examples{
donttest{}
}
