% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_translate.baidu.R
\name{func_translate.baidu}
\alias{func_translate.baidu}
\title{Translate your data by baidu}
\usage{
func_translate.baidu(
  data,
  appid = "20240805002116930",
  key = "s9FjAndS9uQWFIuGiFhh"
)
}
\arguments{
\item{data}{your data.}

\item{appid}{your appid of API.}

\item{key}{your key ou API.}
}
\value{
results of translation.
}
\description{
Translate your data by baidu
}
\examples{
donttest{}
}
