% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_main_script.R
\name{generate_main_script}
\alias{generate_main_script}
\title{Generate main script.}
\usage{
generate_main_script(target_path, overwrite = FALSE)
}
\arguments{
\item{target_path}{Target Generation Path}

\item{overwrite}{Do you want to overwrite?}
}
\value{
target_path
}
\description{
Generate the main R script for this project, which includes the global configuration and basic package installation of this project. It is a necessary script for the start of the project.
}
\examples{
donttest{}
}
